+++ default.nix
@@ -1,11 +1,11 @@
 { system ? null , ... }:
 let
+  version = "18.09";
   # Import a specific Nixpkgs revision to use as the base for our overlay.
-  nixpkgs = (import <nixpkgs> {}).fetchFromGitHub {
-    owner = "NixOS";
-    repo = "nixpkgs";
-    rev = "8c2c14ac392e5b96a1b3d12e16ba0439689024c7";
-    sha256 = "0x3b0ml7gxc9y28y4l64mx6w5582ncks0rca00ikn1pqffffvbxi";
+  nixpkgs = builtins.fetchTarball {
+    name = "nixpkgs-${version}";
+    url = "https://github.com/nixos/nixpkgs/archive/${version}.tar.gz";
+    sha256 = "1ib96has10v5nr6bzf7v8kw7yzww8zanxgw2qi1ll1sbv6kj6zpd";
   };
 in
   # Now return the Nixpkgs configured to use our overlay.