+++ src/hevm/.dir-locals.el
@@ -1,0 +1,1 @@
+((haskell-mode . ((dante-target . "lib:hevm"))))