+++ src/hevm/src/EVM/Op.hs
@@ -67,6 +67,7 @@ data Op
   | OpJumpdest
   | OpCreate
   | OpCall
+  | OpStaticcall
   | OpCallcode
   | OpReturn
   | OpDelegatecall